pyproject.toml
[project]
name = "slk-calendar-sync"
version = "1.0.0"
description = "Automatically sync Super League Kerala match data with Google Calendar"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["football", "kerala", "calendar", "google-calendar", "sports"]
dependencies = [
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic>=2.12.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
]

[project.scripts]
slk-sync = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
